Waypoints view: Split out XBee sender methods
This allows us to reuse the sending progress tracker for other views.
