+++ control_panel/Control_Panel_Waypoints_View.py
@@ -3,6 +3,7 @@ import os
 from functools import partial
 from PyQt4 import QtCore, QtGui
 from Control_Panel_View import Control_Panel_View
+from Control_Panel_XBee_Sender import Control_Panel_XBee_Sender
 from ..zigbee.XBee_Packet import XBee_Packet
 
 class Control_Panel_Waypoints_View(Control_Panel_View):
@@ -11,12 +12,6 @@ class Control_Panel_Waypoints_View(Control_Panel_View):
 
         self._max_retries = self._settings.get("waypoints_max_retries")
         self._retry_interval = self._settings.get("waypoints_retry_interval")
-        self._clear_send()
-
-    def clear(self, layout=None):
-        super(Control_Panel_Waypoints_View, self).clear(layout)
-
-        self._controller.remove_packet_callback("waypoint_ack")
 
     def load(self, data):
         self._listWidget = QtGui.QListWidget()
@@ -65,7 +60,6 @@ class Control_Panel_Waypoints_View(Control_Panel_View):
         """
 
         self._add_menu_bar()
-        self._controller.add_packet_callback("waypoint_ack", self._receive_ack)
 
         self._listWidget.setSizePolicy(QtGui.QSizePolicy.Fixed, QtGui.QSizePolicy.Expanding)
         self._listWidget.setCurrentRow(0)
@@ -257,62 +251,15 @@ class Control_Panel_Waypoints_View(Control_Panel_View):
                                        "There are no vehicles with waypoints.")
             return
 
-        # Create a progress dialog and send the waypoints to the vehicles.
-        self._progress = QtGui.QProgressDialog(self._controller.central_widget)
-        self._progress.setMinimum(0)
-        self._progress.setMaximum(total)
-        self._progress.setWindowModality(QtCore.Qt.WindowModal)
-        self._progress.setMinimumDuration(0)
-        self._progress.setCancelButtonText("Cancel")
-        self._progress.canceled.connect(lambda: self._cancel())
-        self._progress.setWindowTitle("Sending waypoints")
-        self._progress.setLabelText("Initializing...")
-        self._progress.open()
+        sender = Control_Panel_XBee_Sender(self._controller, "waypoint",
+                                           waypoints, total, "waypoint_clear",
+                                           self._make_add_waypoint_packet, 
+                                           "waypoint_done", "waypoint_ack",
+                                           self._max_retries,
+                                           self._retry_interval)
+        sender.start()
 
-        self._labels = dict([(vehicle, "") for vehicle in waypoints])
-
-        self._retry_counts = dict([(vehicle, self._max_retries) for vehicle in waypoints])
-        self._indexes = dict([(vehicle, -1) for vehicle in waypoints])
-
-        self._timers = {}
-        for vehicle in waypoints:
-            timer = QtCore.QTimer()
-            timer.setInterval(self._retry_interval * 1000)
-            timer.setSingleShot(True)
-            # Bind timeout signal to retry for the current vehicle.
-            timer.timeout.connect(partial(self._retry, vehicle))
-            self._timers[vehicle] = timer
-
-        self._waypoints = waypoints
-        self._total = total
-        for vehicle in self._waypoints:
-            self._send_clear(vehicle)
-
-    def _send_clear(self, vehicle):
-        packet = XBee_Packet()
-        packet.set("specification", "waypoint_clear")
-        packet.set("to_id", vehicle)
-
-        self._controller.xbee.enqueue(packet, to=vehicle)
-
-        self._set_label(vehicle, "Clearing old waypoints")
-        self._timers[vehicle].start()
-
-    def _send_one(self, vehicle):
-        index = self._indexes[vehicle]
-        if len(self._waypoints[vehicle]) <= index:
-            # Enqueue a packet indicating that waypoint sending
-            # for this vehicle is done.
-            packet = XBee_Packet()
-            packet.set("specification", "waypoint_done")
-            packet.set("to_id", vehicle)
-            self._controller.xbee.enqueue(packet, to=vehicle)
-
-            self._update_value()
-            return
-
-        waypoint = self._waypoints[vehicle][index]
-
+    def _make_add_waypoint_packet(self, vehicle, index, waypoint):
         packet = XBee_Packet()
         packet.set("specification", "waypoint_add")
         packet.set("latitude", waypoint[0])
@@ -320,72 +267,4 @@ class Control_Panel_Waypoints_View(Control_Panel_View):
         packet.set("index", index)
         packet.set("to_id", vehicle)
 
-        self._controller.xbee.enqueue(packet, to=vehicle)
-
-        self._set_label(vehicle, "Sending waypoint #{} ({}, {})".format(index, waypoint[0], waypoint[1]))
-        self._timers[vehicle].start()
-
-    def _receive_ack(self, packet):
-        vehicle = packet.get("sensor_id")
-        index = packet.get("next_index")
-
-        if vehicle not in self._timers:
-            return
-
-        self._indexes[vehicle] = index
-        self._retry_counts[vehicle] = self._max_retries + 1
-
-    def _retry(self, vehicle):
-        self._retry_counts[vehicle] -= 1
-        if self._retry_counts[vehicle] > 0:
-            if self._indexes[vehicle] == -1:
-                self._send_clear(vehicle)
-            else:
-                self._send_one(vehicle)
-        else:
-            self._cancel("Vehicle {}: Maximum retry attempts for {} reached".format(vehicle, "clearing waypoints" if self._indexes[vehicle] == -1 else "index {}".format(self._indexes[vehicle])))
-
-    def _set_label(self, vehicle, text):
-        self._labels[vehicle] = text
-        self._update_labels()
-        self._update_value()
-
-    def _update_labels(self):
-        labels = []
-        for vehicle in sorted(self._labels.iterkeys()):
-            label = self._labels[vehicle]
-            if self._retry_counts[vehicle] < self._max_retries:
-                retry = " ({} attempts remaining)".format(self._retry_counts[vehicle])
-            else:
-                retry = ""
-
-            labels.append("Vehicle {}: {}{}".format(vehicle, label, retry))
-
-        self._progress.setLabelText("\n".join(labels))
-
-    def _update_value(self):
-        self._progress.setValue(max(0, min(self._total, sum(self._indexes.values()))))
-
-    def _cancel(self, message=None):
-        for timer in self._timers.values():
-            timer.stop()
-
-        if self._progress is not None:
-            self._progress.cancel()
-            self._progress.deleteLater()
-
-        if message is not None:
-            QtGui.QMessageBox.critical(self._controller.central_widget,
-                                       "Sending failed", message)
-
-        self._clear_send()
-
-    def _clear_send(self):
-        # Clear variables used in the _send method and its submethods
-        self._progress = None
-        self._labels = {}
-        self._retry_counts = {}
-        self._indexes = {}
-        self._waypoints = {}
-        self._total = 0
-        self._timers = {}
+        return packet
